Added beginnings of C++ design.
A lot of the existing C code is going to be moved into C++. Anything
that isn't 'leaf' code will be integrated into the object model.
We're going for as much of the dependency injection style as we can,
but OpenGL being a state machine means this will fall apart at the
lower level.