Added the code for drawFrame
Added code for the draw frame function. It finds the position
and orientation of each point on the current frame by
interpolation the two frames it is inbetween. This is done
using the LERP and NLERP function. NLERP can easily be swapped
with SLEPR. It then rotates the quatrenion by the orientation
of its parent and draws it. The code is not commented. It is also
not tested so it is probably broken. But its there (and complete)!