Refactored Md5Model.DrawFrame() a bit.
The code is largely the same, just patched up some omissions and
threw away the notion of a temporary md5AnimFrame as much as was
reasonable. Also added a test model with an animation named
'res/sphere.md5{mesh,anim}'. Finally, it was decided that
frame-rate is more useful than frame duration, so that is what we
store after reading the animation file.