Set a basic template and worked on the animation subsystem
Set up a basic template for the entire document. This includes
three main parts: the introduction, the overall design
and the subsystems. Described the implementation level description
for everything in the animation subsystem excluding the stuff
about shaders, which I know next to nothing about. Yet... Also
added two sections in the appendix, one describing quaternions
and the othet describing interpolation techniques.